Rename `contentlayer` constants file to `docs`
