+++ src/app/layout.tsx
@@ -3,7 +3,7 @@ import { Figtree, Fira_Code } from 'next/font/google';
 import { clsx } from 'clsx';
 
 import Layout from '@/components/layout';
-import { config } from '@/constants/contentlayer';
+import { config } from '@/constants/docs';
 
 import '@/styles/index.css';
 import '@/styles/prism.css';
